Link to correct site
The former domain has been seized by a rogue third party and is NOT affiliated with the Tox project and its developers. 

The website of the Tox project is https://tox.chat. For explanation see https://blog.tox.chat/2015/07/current-situation-3/ and for verification ask around in #tox-dev on freenode.

Thank you!